requirements.txt
wheel==0.24.0
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.1.1
Pygments==2.1.3
appnope==0.1.0
backports-abc==0.4
cycler==0.10.0
dask==0.8.1
decorator==4.0.9
gnureadline==6.3.3
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
mahotas==1.4.1
matplotlib==1.5.1
mistune==0.7.2
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
notebook==4.1.0
numpy==1.10.4
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
git+https://github.com/darcymason/pydicom/
pyparsing==2.1.0
python-dateutil==2.5.0
pytz==2016.1
pyzmq==15.2.0
qtconsole==4.2.0
scikit-image==0.12.3
scipy==0.17.0
simplegeneric==0.8.1
six==1.10.0
stevedore==1.12.0
terminado==0.6
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
